Extract seed/reference/knowledge MCP tools out of mcp-server.ts
Continues Wave C (batch 4, mcp-server.ts 4817 -> 4578 lines). Moves
get_reference_images/get_seed/regenerate_seed/knowledge_query into
mcp-seed-tools.ts as createSeedTools({resolveAuth, injectChannel}).
All four are verbatim moves with the established thin-proxy pattern
from batches 1-3 -- no self-references, no exclusive private helpers.
