+++ eslint-rules/module-boundary-map.mjs
@@ -180,6 +180,7 @@ const FILE_BUCKET = {
   'mcp-kit-tools': 'agent-surface',
   'mcp-kit-file-tools': 'agent-surface',
   'mcp-inbox-tools': 'agent-surface',
+  'mcp-seed-tools': 'agent-surface',
   'mcp-server-discovery': 'agent-surface',
   'mcp-ui': 'agent-surface',
   'mcp-oauth-metadata': 'agent-surface',
